Add interfaces to ListAllPacks
